+++ core-data/src/commonMain/kotlin/ai/saniou/thread/data/source/tieba/TiebaSource.kt
@@ -49,9 +49,11 @@ class TiebaSource(
         supportsPagination = true,
         hasSubComments = true,
         hasUpvote = true,
+        hasDownvote = true,
         hasPoOnly = true,
         supportsChannelSign = true,
         supportsForumRules = true,
+        supportsUserFollow = true,
     )
 
     override fun topicUrl(topicId: String): String = "https://tieba.baidu.com/p/$topicId"