+++ feature-forum/src/commonMain/kotlin/ai/saniou/forum/workflow/topicdetail/TopicDetailContract.kt
@@ -109,6 +109,11 @@ interface TopicDetailContract {
         object UpvoteTopic : Event
 
         /**
+         * 点踩主贴
+         */
+        object DownvoteTopic : Event
+
+        /**
          * 显示楼中楼弹窗
          * @param commentId 评论ID
          */